Hoist password guesser out of TextSlide, render guesser on phone
